Handle metadata parsing in the Trigger classes
