+++ pytroll_collectors/geographic_gatherer.py
@@ -27,7 +27,6 @@
 import logging
 import time
 import datetime as dt
-import os.path
 
 from six.moves.configparser import NoOptionError
 from posttroll import publisher
@@ -35,7 +34,7 @@ from satpy.resample import get_area_def
 from trollsift import Parser
 
 from pytroll_collectors.region_collector import RegionCollector
-from pytroll_collectors.trigger import fix_start_end_time, PostTrollTrigger, WatchDogTrigger, terminator_function
+from pytroll_collectors.trigger import PostTrollTrigger, WatchDogTrigger, terminator_function
 
 logger = logging.getLogger(__name__)
 
@@ -76,7 +75,6 @@ class GeographicGatherer(object):
             trigger.start()
             self.triggers.append(trigger)
 
-
     def _get_collectors(self, section, regions):
         timeliness = dt.timedelta(minutes=self._config.getint(section, "timeliness"))
         try:
@@ -108,37 +106,11 @@ class GeographicGatherer(object):
         return WatchDogTrigger(
             collectors,
             terminator_function,
-            self._get_metadata,
+            self._config,
             [glob],
             observer_class,
             self.publisher,
             publish_topic=publish_topic)
-
-    def _get_metadata(self, fname):
-        """Parse metadata from the file."""
-        res = None
-        for section in self._config.sections():
-            try:
-                parser = Parser(self._config.get(section, "pattern"))
-            except NoOptionError:
-                continue
-            if not parser.validate(fname):
-                continue
-            res = parser.parse(fname)
-            res.update(dict(self._config.items(section)))
-
-            for key in ["watcher", "pattern", "timeliness", "regions"]:
-                res.pop(key, None)
-
-            res = fix_start_end_time(res)
-
-            if ("sensor" in res) and ("," in res["sensor"]):
-                res["sensor"] = res["sensor"].split(",")
-
-            res["uri"] = fname
-            res["filename"] = os.path.basename(fname)
-
-        return res
 
     def _get_publish_topic(self, section):
         try:
@@ -146,7 +118,6 @@ class GeographicGatherer(object):
         except NoOptionError:
             publish_topic = None
         return publish_topic
-
 
     def _get_posttroll_trigger(self, section, observer_class, collectors):
         logger.debug("Using posttroll for %s", section)
@@ -164,14 +135,12 @@ class GeographicGatherer(object):
             publish_topic=publish_topic, nameserver=nameserver,
             publish_message_after_each_reception=publish_message_after_each_reception)
 
-
     def _get_nameserver(self, section):
         try:
             nameserver = self._config.get(section, "nameserver")
         except NoOptionError:
             nameserver = "localhost"
         return nameserver
-
 
     def _get_duration(self, section):
         try:
